Changed not to locate other symbols (-s) in GTAGSLIBPATH.
Here is the RFC:

Subject: [RFC] Behavior of global(1) when the GTAGSLIBPATH is set.
Date: Tue, 22 Aug 2006 10:27:12 +0900
Sender: shigio@tamacom.com

Hello all,
I would like to change the behavior of global(1) when the
GTAGSLIBPATH is set.

When the GTAGSLIBPATH variable is set, global(1) retrieve in not
only current project but also library projects listed in the variable.
At present, this facility works for not only definition retrieval
but also other symbol retrieval.

[PRESENT SPEC]
option   retrieval type  retrieve in GTAGSLIBPATH?
-------------------------------------------------
(none)   definition      yes
-r       reference       no
-s       other symbol    yes <=

But I think this specification is wrong.
Global(1) should retrieve only the objects clearly defined
in the library projects; it should not retrieve other symbols.
Because the library is these kind of things; it is hardly thought
to refer to the symbols in the library modules.

In the past, there was a request that want to retrieve the symbols
(-s) in the library projects. I accepted the request, and changed
the specification like present, since there was no method to
retrieve them through multiple projects besides GTAGSLIBPATH.
However, since gtags has the --file option now, and we can bring
multiple projects together with it, it is not necessary to support
such ugly facility.

Thus, I would like to change the specification like follows:

[NEW SPEC(=original)]
option   retrieval type  retrieve in GTAGSLIBPATH?
-------------------------------------------------
(none)   definition      yes
-r       reference       no
-s       other symbol    *no*

What do you think?
Is there something problem?